+++ WomxnDevelopUbisoftDemo/Game/ObjectsElements.cpp
@@ -11,7 +11,8 @@ ObjectsElements::ObjectsElements(sf::Vector2f& spawn_pos, bool canmove, bool ani
 
 	// Set texture 
 	e_Sprite.setTexture(*m_pTextureAtlas);
-	e_Sprite.setTextureRect(sf::IntRect(upperleft.x, upperleft.y, e_size.x, e_size.y));
+	e_Sprite.setTextureRect(sf::IntRect(upperleft.x, upperleft.y, 
+							static_cast<int>(e_size.x), static_cast<int>(e_size.y)));
 	// Set Origin
 	e_Sprite.setOrigin(e_size * 0.5f);
 	// Set Position
